Resolve "controller skelekton code"

+++ rover/Wall-E/lib/include/pin_assignment.h
@@ -1,13 +1,12 @@
 #pragma once
-#include <cstdint>
-
+#include <Arduino.h>
 namespace pin
 {
     /**
-     * @brief 
+     * @brief
      *
      */
-    const static uint8_t TEST_DUMMPY = 255;
-    const static uint8_t MOTOR_PWM_1 = 1;
-    const static uint8_t RC_INPUT    = 2;
+    const static int TEST_DUMMPY = 255;
+    const static int MOTOR_PWM_1 = 1;
+    const static int RC_INPUT    = 2;
 }  // namespace pin
